adding types and reorganize the structure
